testes/requirements.txt
alembic==1.7.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==3.2.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-resources==5.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.0
python-dotenv==0.20.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.5.0
